requirements.txt
tensorflow
numpy
hydra-core
